chore(ci): detectChanges: Check for "code" changes (#10390)
