+++ .github/workflows/check-test-project-fixture.yml
@@ -17,7 +17,7 @@ jobs:
     runs-on: ubuntu-latest
 
     outputs:
-      docs_only: ${{ steps.detect-changes.outputs.docs_only }}
+      code: ${{ steps.detect-changes.outputs.code }}
       rsc: ${{ steps.detect-changes.outputs.rsc }}
       ssr: ${{ steps.detect-changes.outputs.ssr }}
 
@@ -36,7 +36,7 @@ jobs:
 
   check-test-project-fixture:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'false'
+    if: needs.detect-changes.outputs.code == 'true'
     name: Check test project fixture
     runs-on: ubuntu-latest
     steps:
@@ -74,7 +74,7 @@ jobs:
 
   check-test-project-fixture-skip:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'true'
+    if: needs.detect-changes.outputs.code == 'false'
     name: Check test project fixture
     runs-on: ubuntu-latest
     steps:
