+++ .github/workflows/ci.yml
@@ -22,8 +22,7 @@ jobs:
     runs-on: ubuntu-latest
 
     outputs:
-      docs_only: ${{ steps.detect-changes.outputs.docs_only }}
-      changesets_only: ${{ steps.detect-changes.outputs.changesets_only }}
+      code: ${{ steps.detect-changes.outputs.code }}
       rsc: ${{ steps.detect-changes.outputs.rsc }}
       ssr: ${{ steps.detect-changes.outputs.ssr }}
 
@@ -42,7 +41,7 @@ jobs:
 
   check:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'false' && needs.detect-changes.outputs.changesets_only == 'false'
+    if: needs.detect-changes.outputs.code == 'true'
 
     name: ✅ Check constraints, dependencies, and package.json's
     runs-on: ubuntu-latest
@@ -61,7 +60,7 @@ jobs:
 
   check-skip:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'true' || needs.detect-changes.outputs.changesets_only == 'true'
+    if: needs.detect-changes.outputs.code == 'false'
 
     name: ✅ Check constraints, dependencies, and package.json's
     runs-on: ubuntu-latest
@@ -104,7 +103,7 @@ jobs:
 
   build-lint-test-skip:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'true'
+    if: needs.detect-changes.outputs.code == 'false'
 
     strategy:
       matrix:
@@ -178,7 +177,7 @@ jobs:
 
   tutorial-e2e-skip:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'true'
+    if: needs.detect-changes.outputs.code == 'false'
 
     strategy:
       matrix:
@@ -269,7 +268,7 @@ jobs:
 
   smoke-tests-skip:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'true'
+    if: needs.detect-changes.outputs.code == 'false'
 
     strategy:
       matrix:
@@ -379,7 +378,7 @@ jobs:
 
   cli-smoke-tests-skip:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'true'
+    if: needs.detect-changes.outputs.code == 'false'
 
     strategy:
       matrix:
@@ -419,7 +418,7 @@ jobs:
 
   telemetry-check-skip:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'true'
+    if: needs.detect-changes.outputs.code == 'false'
 
     strategy:
       matrix:
@@ -751,7 +750,7 @@ jobs:
 
   crwa-skip:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'true'
+    if: needs.detect-changes.outputs.code == 'false'
 
     name: 🌲 Create Redwood App
     runs-on: ubuntu-latest
@@ -775,7 +774,7 @@ jobs:
 
   server-tests-skip:
     needs: detect-changes
-    if: needs.detect-changes.outputs.docs_only == 'true'
+    if: needs.detect-changes.outputs.code == 'false'
 
     name: Server tests
     runs-on: ubuntu-latest
